app/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
customidenticon==0.1.5
cycler==0.10.0
descartes==1.1.0
Flask==1.0.2
Flask-QRcode==2.0.2
Flask-SocketIO==4.2.0
Flask-Stache==0.1.1
graphviz==0.14
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.3
mizani==0.5.4
mysql-connector==2.2.9
numpy==1.18.5
palettable==3.3.0
pandas==0.24.2
patsy
plotnine==0.5.1
pyparsing==2.4.7
PySocks==1.7.1
pystache==0.5.4
python-dateutil==2.8.1
python-engineio==3.13.0
python-socketio==4.6.0
pytz==2020.1
qrcode==6.1
requests==2.20.0
scipy==1.4.1
six==1.12.0
statsmodels==0.11.1
urllib3==1.24.2
Werkzeug==0.15.2
